Split label tests from 1 file to 4 files.
