--- project/labels/tests/test_labels.py
@@ -1,1117 +1,0 @@
-from __future__ import unicode_literals
-
-from django.contrib.auth import get_user_model
-from django.contrib.auth.models import Group
-from django.core.cache import cache
-from django.core import mail
-from django.shortcuts import resolve_url
-from django.test import override_settings
-from django.urls import reverse
-from django.utils.html import escape as html_escape
-
-from images.model_utils import PointGen
-from images.models import Source
-from lib.tests.utils import (
-    BasePermissionTest, ClientTest, sample_image_as_file)
-from ..models import LabelGroup, Label
-from ..templatetags.labels import (
-    popularity_bar as popularity_bar_tag, status_icon as status_icon_tag)
-from .utils import LabelTest
-
-User = get_user_model()
-
-
-class PermissionTest(BasePermissionTest):
-    """
-    Test permissions of label views other than label edit, which is tested in
-    another class.
-    """
-    @classmethod
-    def setUpTestData(cls):
-        super(PermissionTest, cls).setUpTestData()
-
-        cls.labels = cls.create_labels(cls.user, ['A', 'B'], 'GroupA')
-
-    def test_label_list(self):
-        url = reverse('label_list')
-        template = 'labels/label_list.html'
-
-        self.assertPermissionLevel(url, self.SIGNED_OUT, template=template)
-
-    def test_label_list_search_ajax(self):
-        url = reverse('label_list_search_ajax')
-
-        self.assertPermissionLevel(url, self.SIGNED_OUT, is_json=True)
-
-    def test_label_new(self):
-        url = reverse('label_new')
-        template = 'labels/label_new.html'
-
-        self.assertPermissionLevel(
-            url, self.SIGNED_IN, template=template,
-            deny_type=self.REQUIRE_LOGIN)
-
-    def test_label_new_ajax(self):
-        url = reverse('label_new_ajax')
-
-        self.assertPermissionLevel(
-            url, self.SIGNED_IN, is_json=True, post_data={},
-            deny_type=self.REQUIRE_LOGIN)
-
-    def test_label_main(self):
-        url = reverse('label_main', args=[self.labels[0].pk])
-        template = 'labels/label_main.html'
-
-        self.assertPermissionLevel(url, self.SIGNED_OUT, template=template)
-
-    def test_label_example_patches_ajax(self):
-        url = reverse('label_example_patches_ajax', args=[self.labels[0].pk])
-
-        self.assertPermissionLevel(url, self.SIGNED_OUT, is_json=True)
-
-
-class LabelListTest(ClientTest):
-    """
-    Test the label list page.
-    """
-    @classmethod
-    def setUpTestData(cls):
-        # Call the parent's setup (while still using this class as cls)
-        super(LabelListTest, cls).setUpTestData()
-
-        cls.user = cls.create_user()
-
-        cls.user_committee_member = cls.create_user()
-        cls.user_committee_member.groups.add(
-            Group.objects.get(name="Labelset Committee"))
-
-        # Create labels
-        cls.labels = cls.create_labels(
-            cls.user, ['A', 'B'], "Group1")
-
-    def test_load_page(self):
-        response = self.client.get(reverse('label_list'))
-        self.assertStatusOK(response)
-
-    def test_new_label_link_not_shown_for_regular_user(self):
-        self.client.force_login(self.user)
-        response = self.client.get(reverse('label_list'))
-        self.assertNotContains(response, "Create a new label")
-
-    def test_new_label_link_shown_for_committee_member(self):
-        self.client.force_login(self.user_committee_member)
-        response = self.client.get(reverse('label_list'))
-        self.assertContains(response, "Create a new label")
-
-
-class BaseLabelSearchTest(ClientTest):
-
-    def assertLabels(self, response, label_names):
-        response_pk_set = set(response.json()['label_ids'])
-        expected_pk_set = set(Label.objects.filter(
-            name__in=label_names).values_list('pk', flat=True))
-        self.assertSetEqual(response_pk_set, expected_pk_set)
-
-    def submit_search(self, **kwargs):
-        data = dict(
-            name_search='',
-            show_verified=True,
-            show_regular=True,
-            show_duplicate=False,
-            functional_group='',
-            min_popularity='',
-        )
-        data.update(kwargs)
-        response = self.client.get(self.url, data)
-        return response
-
-    def setUp(self):
-        super(BaseLabelSearchTest, self).setUp()
-
-        # Popularities are cached when computed, so we clear the cache to
-        # prevent a previous test from affecting the next one.
-        cache.clear()
-
-
-class LabelSearchNameFieldTest(BaseLabelSearchTest):
-    """
-    Test the name field of the Ajax label search.
-    """
-    @classmethod
-    def setUpTestData(cls):
-        super(LabelSearchNameFieldTest, cls).setUpTestData()
-
-        cls.user = cls.create_user()
-        cls.url = reverse('label_list_search_ajax')
-
-    def test_match_full_name(self):
-        self.create_labels(self.user, ["Red", "Blue"], "Group1")
-
-        response = self.submit_search(name_search="Red")
-        self.assertLabels(response, ["Red"])
-
-    def test_match_part_of_name(self):
-        self.create_labels(self.user, ["Red", "Blue"], "Group1")
-
-        response = self.submit_search(name_search="Blu")
-        self.assertLabels(response, ["Blue"])
-
-    def test_match_case_insensitive(self):
-        self.create_labels(self.user, ["Red", "Blue"], "Group1")
-
-        response = self.submit_search(name_search="BLUE")
-        self.assertLabels(response, ["Blue"])
-
-    def test_match_multiple_labels(self):
-        self.create_labels(
-            self.user, ["Red", "Light Blue", "Dark Blue"], "Group1")
-
-        response = self.submit_search(name_search="Blue")
-        self.assertLabels(response, ["Light Blue", "Dark Blue"])
-
-    def test_multiple_words(self):
-        self.create_labels(
-            self.user, ["Light Blue", "Dark Blue", "Dark Red"], "Group1")
-
-        response = self.submit_search(name_search="Dark Blue")
-        self.assertLabels(response, ["Dark Blue"])
-
-    def test_no_match(self):
-        self.create_labels(self.user, ["Red", "Blue"], "Group1")
-
-        response = self.submit_search(name_search="Green")
-        self.assertLabels(response, [])
-
-    def test_strip_whitespace(self):
-        self.create_labels(self.user, ["Blue", "Red"], "Group1")
-
-        response = self.submit_search(name_search="  Blue ")
-        self.assertLabels(response, ["Blue"])
-
-    def test_normalize_multiple_spaces(self):
-        self.create_labels(
-            self.user, ["Light Blue", "Dark Blue", "Dark Red"], "Group1")
-
-        response = self.submit_search(name_search="Dark   Blue")
-        self.assertLabels(response, ["Dark Blue"])
-
-    def test_treat_punctuation_as_spaces(self):
-        self.create_labels(
-            self.user, ["Light Blue", "Dark Blue", "Dark Red"], "Group1")
-
-        response = self.submit_search(name_search=";'Dark_/Blue=-")
-        self.assertLabels(response, ["Dark Blue"])
-
-    def test_no_tokens(self):
-        self.create_labels(
-            self.user, ["Light Blue", "Dark Blue", "Dark Red"], "Group1")
-
-        response = self.submit_search(name_search=";'_/=-")
-        self.assertLabels(response, [])
-
-
-class LabelSearchOtherFieldsTest(BaseLabelSearchTest):
-    """Test fields other than name."""
-    @classmethod
-    def setUpTestData(cls):
-        super(LabelSearchOtherFieldsTest, cls).setUpTestData()
-
-        cls.user = cls.create_user()
-        cls.url = reverse('label_list_search_ajax')
-
-        cls.create_labels(cls.user, ['A', 'B'], "Group1")
-        cls.create_labels(cls.user, ['C', 'D'], "Group2")
-
-    def test_show_by_status(self):
-        label_a = Label.objects.get(name='A')
-        label_a.verified = True
-        label_a.save()
-        label_b = Label.objects.get(name='B')
-        label_b.duplicate = label_a
-        label_b.save()
-
-        response = self.submit_search(
-            show_verified=False, show_regular=False, show_duplicate=False)
-        self.assertLabels(response, [])
-
-        response = self.submit_search(
-            show_verified=True, show_regular=False, show_duplicate=False)
-        self.assertLabels(response, ['A'])
-
-        response = self.submit_search(
-            show_verified=False, show_regular=True, show_duplicate=False)
-        self.assertLabels(response, ['C', 'D'])
-
-        response = self.submit_search(
-            show_verified=False, show_regular=False, show_duplicate=True)
-        self.assertLabels(response, ['B'])
-
-        response = self.submit_search(
-            show_verified=True, show_regular=True, show_duplicate=False)
-        self.assertLabels(response, ['A', 'C', 'D'])
-
-        response = self.submit_search(
-            show_verified=True, show_regular=True, show_duplicate=True)
-        self.assertLabels(response, ['A', 'B', 'C', 'D'])
-
-    def test_show_by_functional_group(self):
-        response = self.submit_search()
-        self.assertLabels(response, ['A', 'B', 'C', 'D'])
-
-        response = self.submit_search(
-            functional_group=LabelGroup.objects.get(name='Group1').pk)
-        self.assertLabels(response, ['A', 'B'])
-
-        response = self.submit_search(
-            functional_group=LabelGroup.objects.get(name='Group2').pk)
-        self.assertLabels(response, ['C', 'D'])
-
-    def test_show_by_popularity(self):
-        source = self.create_source(
-            self.user,
-            point_generation_type=PointGen.Types.SIMPLE,
-            simple_number_of_points=4)
-        self.create_labelset(self.user, source, Label.objects.all())
-        img = self.upload_image(self.user, source)
-        self.add_annotations(
-            self.user, img, {1: 'A', 2: 'A', 3: 'C', 4: 'C'})
-
-        response = self.submit_search()
-        self.assertLabels(response, ['A', 'B', 'C', 'D'])
-
-        response = self.submit_search(min_popularity=1)
-        self.assertLabels(response, ['A', 'C'])
-
-    def test_multiple_filters(self):
-        label_a = Label.objects.get(name='A')
-        label_b = Label.objects.get(name='B')
-        label_c = Label.objects.get(name='C')
-
-        # A, B, and C are verified
-        label_a.verified = True
-        label_a.save()
-        label_b.verified = True
-        label_b.save()
-        label_c.verified = True
-        label_c.save()
-
-        # A and B are already in Group1
-
-        # A and C have popularity
-        source = self.create_source(
-            self.user,
-            point_generation_type=PointGen.Types.SIMPLE,
-            simple_number_of_points=4)
-        self.create_labelset(self.user, source, Label.objects.all())
-        img = self.upload_image(self.user, source)
-        self.add_annotations(
-            self.user, img, {1: 'A', 2: 'A', 3: 'C', 4: 'C'})
-
-        # Only A satisfies all requirements here
-        response = self.submit_search(
-            show_verified=True, show_regular=False,
-            functional_group=LabelGroup.objects.get(name='Group1').pk,
-            min_popularity=1)
-        self.assertLabels(response, ['A'])
-
-
-class LabelDetailTest(ClientTest):
-    """
-    Test the label detail page.
-    """
-    @classmethod
-    def setUpTestData(cls):
-        super(LabelDetailTest, cls).setUpTestData()
-
-        cls.user = cls.create_user()
-
-        cls.group = LabelGroup(name="Group 1", code='G1')
-        cls.group.save()
-
-    def test_basic_fields(self):
-        label = Label(
-            name="Label A",
-            default_code='A',
-            group=self.group,
-            description="This is a\nmultiline description.",
-            # This filename will be discarded in favor of a generated one.
-            thumbnail=sample_image_as_file('_.png'),
-            created_by=self.user,
-        )
-        label.save()
-
-        response = self.client.get(reverse('label_main', args=[label.pk]))
-        self.assertStatusOK(response)
-
-        self.assertContains(response, "Name: Label A")
-
-        self.assertContains(response, "Functional Group: Group 1")
-
-        self.assertContains(response, "Default Short Code: A")
-
-        self.assertInHTML(
-            "This is a<br>multiline description.", response.content)
-
-        self.assertInHTML(
-            '<img src="{}" alt="Label A" class="label-thumbnail">'.format(
-                label.thumbnail.url),
-            response.content)
-
-        # Too lazy to check the date itself, but there should be a line for it.
-        self.assertContains(response, "Create Date:")
-
-        self.assertContains(
-            response, "Created By: {}".format(self.user.username))
-
-    def test_duplicate(self):
-        labels = self.create_labels(self.user, ['A', 'B'], "Group1")
-
-        # Non-duplicate
-        label_a = labels.get(name='A')
-        response = self.client.get(reverse('label_main', args=[label_a.pk]))
-        self.assertNotContains(response, "THIS LABEL IS A DUPLICATE OF")
-
-        label_b = labels.get(name='B')
-        label_b.verified = True
-        label_b.save()
-        label_a.duplicate = label_b
-        label_a.save()
-
-        # Duplicate
-        response = self.client.get(reverse('label_main', args=[label_a.pk]))
-        self.assertContains(response, "THIS LABEL IS A DUPLICATE OF")
-        self.assertInHTML(
-            'THIS LABEL IS A DUPLICATE OF: <a href="{}">B</a>'.format(
-                reverse('label_main', args=[label_b.pk])
-            ),
-            response.content)
-
-    def test_verified(self):
-        labels = self.create_labels(self.user, ['A'], "Group1")
-        label_a = labels.get(name='A')
-
-        response = self.client.get(reverse('label_main', args=[label_a.pk]))
-        self.assertInHTML("Verified: No", response.content)
-
-        label_a.verified = True
-        label_a.save()
-
-        response = self.client.get(reverse('label_main', args=[label_a.pk]))
-
-        status_icon_html = status_icon_tag(label_a)
-        verified_html = 'Verified: Yes {}'.format(status_icon_html)
-        self.assertInHTML(verified_html, response.content)
-
-    def test_usage_info(self):
-        labels = self.create_labels(self.user, ['A', 'B'], "Group1")
-        label_a = labels.get(name='A')
-
-        user_2 = self.create_user()
-
-        user_private_s = self.create_source(
-            self.user, visibility=Source.VisibilityTypes.PRIVATE,
-            name="User's private source")
-        self.create_labelset(self.user, user_private_s, labels)
-        img = self.upload_image(self.user, user_private_s)
-        self.add_annotations(self.user, img, {1: 'A'})
-
-        # No annotation, but has A in the labelset
-        user2_public_s = self.create_source(
-            user_2, visibility=Source.VisibilityTypes.PUBLIC,
-            name="User 2's public source")
-        self.create_labelset(user_2, user2_public_s, labels)
-
-        # Doesn't have A in the labelset
-        user_other_s = self.create_source(
-            self.user, name="User's other source")
-        self.create_labelset(self.user, user_other_s, labels.filter(name='B'))
-        img = self.upload_image(self.user, user_other_s)
-        self.add_annotations(self.user, img, {1: 'B'})
-
-        self.client.force_login(self.user)
-        response = self.client.get(reverse('label_main', args=[label_a.pk]))
-
-        # Usage stats.
-        self.assertInHTML(
-            'Stats: Used in 2 sources and for 1 annotations', response.content)
-
-        # Sources using the label.
-        # Viewer's private sources first, with strong links.
-        # Then other public sources, with links.
-        # (Then other private sources, without links... but these need to have
-        # at least 100 images to be listed, so we won't bother testing that
-        # here unless we make that threshold flexible.)
-        self.assertInHTML(
-            '<a href="{}"><strong>{}</strong></a> |'
-            ' <a href="{}">{}</a> |'.format(
-                reverse('source_main', args=[user_private_s.pk]),
-                html_escape("User's private source"),
-                reverse('source_main', args=[user2_public_s.pk]),
-                html_escape("User 2's public source")),
-            response.content)
-
-        # Popularity.
-        popularity_str = str(int(label_a.popularity)) + '%'
-        popularity_bar_html = popularity_bar_tag(label_a)
-        self.assertInHTML(
-            'Popularity: {} {}'.format(
-                popularity_str, popularity_bar_html),
-            response.content)
-
-
-class LabelDetailPatchesTest(ClientTest):
-    """
-    Test the example annotation patches used by the label detail page.
-    """
-    @classmethod
-    def setUpTestData(cls):
-        # Call the parent's setup (while still using this class as cls)
-        super(LabelDetailPatchesTest, cls).setUpTestData()
-
-        cls.user = cls.create_user()
-        cls.source = cls.create_source(
-            cls.user,
-            point_generation_type=PointGen.Types.SIMPLE,
-            simple_number_of_points=100,
-        )
-
-        cls.labels = cls.create_labels(
-            cls.user, ['A', 'B'], "Group1")
-        cls.create_labelset(cls.user, cls.source, cls.labels)
-        cls.source.refresh_from_db()
-
-        cls.img = cls.upload_image(cls.user, cls.source)
-
-    def test_one_page_of_patches(self):
-        annotations = {1: 'A', 2: 'A', 3: 'A', 4: 'B', 5: 'B'}
-        self.add_annotations(self.user, self.img, annotations)
-
-        response = self.client.get(reverse(
-            'label_example_patches_ajax',
-            args=[Label.objects.get(name='A').id])).json()
-
-        # 3 patch images
-        self.assertEqual(response['patchesHtml'].count('<img'), 3)
-        # Is the last page of patches
-        self.assertEqual(response['isLastPage'], True)
-
-    def test_multiple_pages_of_patches(self):
-        annotations = dict()
-        for n in range(1, 10+1):
-            annotations[n] = 'B'
-        for n in range(11, 63+1):
-            annotations[n] = 'A'
-        self.add_annotations(self.user, self.img, annotations)
-
-        # Page 1: 50 patch images
-        response = self.client.get(reverse(
-            'label_example_patches_ajax',
-            args=[Label.objects.get(name='A').id])).json()
-        self.assertEqual(response['patchesHtml'].count('<img'), 50)
-        self.assertEqual(response['isLastPage'], False)
-
-        # Page 2: 3 patch images
-        response = self.client.get(
-            reverse(
-                'label_example_patches_ajax',
-                args=[Label.objects.get(name='A').id]),
-            dict(page=2),
-        ).json()
-        self.assertEqual(response['patchesHtml'].count('<img'), 3)
-        self.assertEqual(response['isLastPage'], True)
-
-    def test_zero_patches(self):
-        annotations = {1: 'B', 2: 'B'}
-        self.add_annotations(self.user, self.img, annotations)
-
-        response = self.client.get(reverse(
-            'label_example_patches_ajax',
-            args=[Label.objects.get(name='A').id])).json()
-
-        self.assertEqual(response['patchesHtml'].count('<img'), 0)
-        self.assertEqual(response['isLastPage'], True)
-
-
-class LabelDetailPatchLinksTest(ClientTest):
-    """
-    Test the links on the annotation patches.
-    """
-    @classmethod
-    def setUpTestData(cls):
-        # Call the parent's setup (while still using this class as cls)
-        super(LabelDetailPatchLinksTest, cls).setUpTestData()
-
-        cls.user = cls.create_user()
-        cls.users_private_source = cls.create_source(
-            cls.user,
-            visibility=Source.VisibilityTypes.PRIVATE,
-            point_generation_type=PointGen.Types.SIMPLE,
-            simple_number_of_points=5,
-        )
-
-        cls.user2 = cls.create_user()
-        cls.public_source = cls.create_source(
-            cls.user2,
-            visibility=Source.VisibilityTypes.PUBLIC,
-            point_generation_type=PointGen.Types.SIMPLE,
-            simple_number_of_points=5,
-        )
-        cls.other_private_source = cls.create_source(
-            cls.user2,
-            visibility=Source.VisibilityTypes.PRIVATE,
-            point_generation_type=PointGen.Types.SIMPLE,
-            simple_number_of_points=5,
-        )
-
-        # Create labels
-        cls.labels = cls.create_labels(
-            cls.user, ['A', 'B'], "Group1")
-
-        # Add all labels to each source's labelset
-        cls.create_labelset(cls.user2, cls.public_source, cls.labels)
-        cls.public_source.refresh_from_db()
-        cls.create_labelset(cls.user, cls.users_private_source, cls.labels)
-        cls.users_private_source.refresh_from_db()
-        cls.create_labelset(cls.user2, cls.other_private_source, cls.labels)
-        cls.other_private_source.refresh_from_db()
-
-        # Upload an image to each source
-        cls.public_img = cls.upload_image(cls.user2, cls.public_source)
-        cls.users_private_img = cls.upload_image(
-            cls.user, cls.users_private_source)
-        cls.other_private_img = cls.upload_image(
-            cls.user2, cls.other_private_source)
-
-    def test_dont_link_to_others_private_images(self):
-        annotations = {1: 'A', 2: 'A', 3: 'A', 4: 'A'}
-        self.add_annotations(self.user2, self.public_img, annotations)
-        annotations = {1: 'A', 2: 'A'}
-        self.add_annotations(self.user, self.users_private_img, annotations)
-        annotations = {1: 'A'}
-        self.add_annotations(self.user2, self.other_private_img, annotations)
-
-        self.client.force_login(self.user)
-        response = self.client.get(reverse(
-            'label_example_patches_ajax',
-            args=[Label.objects.get(name='A').id])).json()
-
-        # Patches shown: 4 + 2 + 1
-        self.assertEqual(response['patchesHtml'].count('<img'), 7)
-        # Patches with links: 4 + 2
-        self.assertEqual(response['patchesHtml'].count('<a'), 6)
-
-
-class NewLabelAjaxTest(ClientTest):
-    """
-    Test label creation.
-    """
-    @classmethod
-    def setUpTestData(cls):
-        # Call the parent's setup (while still using this class as cls)
-        super(NewLabelAjaxTest, cls).setUpTestData()
-
-        cls.user = cls.create_user()
-
-        # Create labels and group
-        cls.create_labels(
-            cls.user, ['A', 'B'], "Group1")
-
-        cls.url = reverse('label_new_ajax')
-
-    @override_settings(LABELSET_COMMITTEE_EMAIL='labels@example.com')
-    def test_label_creation(self):
-        """Successfully create a new label."""
-        self.client.force_login(self.user)
-        self.client.post(self.url, dict(
-            name="CCC",
-            default_code='C',
-            group=LabelGroup.objects.get(code='Group1').pk,
-            description="Species C.",
-            # A new filename will be generated, and this uploaded
-            # filename will be discarded, so it doesn't matter.
-            thumbnail=sample_image_as_file('_.png'),
-        ))
-
-        # Check that the label was created, and has the expected field values
-        label = Label.objects.get(name="CCC")
-        self.assertEqual(label.default_code, 'C')
-        self.assertEqual(label.group.code, 'Group1')
-        self.assertEqual(label.description, "Species C.")
-        self.assertIsNotNone(label.thumbnail)
-        self.assertEqual(label.verified, False)
-        self.assertEqual(label.created_by_id, self.user.pk)
-
-        # Check that an email was sent out
-        self.assertEqual(len(mail.outbox), 1)
-        label_email = mail.outbox[-1]
-        # Check recipients
-        self.assertEqual(len(label_email.to), 1)
-        self.assertEqual(label_email.to[0], 'labels@example.com')
-        self.assertEqual(len(label_email.cc), 1)
-        self.assertEqual(label_email.cc[0], self.user.email)
-        # Contents should have the creating user, label name, and link
-        # to label details
-        self.assertIn(self.user.username, label_email.body)
-        self.assertIn("CCC", label_email.body)
-        self.assertIn(resolve_url('label_main', label.pk), label_email.body)
-
-    def test_label_name_taken(self):
-        """Name taken -> error."""
-        self.client.force_login(self.user)
-        response = self.client.post(self.url, dict(
-            name="B",
-            default_code='B2',
-            group=LabelGroup.objects.get(code='Group1').pk,
-            description="Species B.",
-            thumbnail=sample_image_as_file('_.png'),
-        ))
-
-        self.assertEqual(response.json(), dict(error=(
-            'Name: There is already a label with the same name:'
-            ' <a href="{url}" target="_blank">B</a>').format(
-                url=reverse(
-                    'label_main', args=[Label.objects.get(name="B").pk]))
-        ))
-
-    def test_cant_verify(self):
-        """Can't set label as verified upon creation."""
-        self.client.force_login(self.user)
-        self.client.post(self.url, dict(
-            name="Label C",
-            default_code='C',
-            group=LabelGroup.objects.get(code='Group1').pk,
-            description="Species C.",
-            thumbnail=sample_image_as_file('_.png'),
-            verified=True,
-        ))
-
-        label = Label.objects.get(name="Label C")
-        self.assertEqual(label.verified, False)
-
-    def test_thumbnail_scaled_down(self):
-        self.client.force_login(self.user)
-        self.client.post(self.url, dict(
-            name="Label C",
-            default_code='C',
-            group=LabelGroup.objects.get(code='Group1').pk,
-            description="Species C.",
-            thumbnail=sample_image_as_file('_.png', image_options=dict(
-                width=Label.THUMBNAIL_WIDTH+100,
-                height=Label.THUMBNAIL_HEIGHT+150)),
-        ))
-
-        label = Label.objects.get(name="Label C")
-        self.assertLessEqual(label.thumbnail.width, Label.THUMBNAIL_WIDTH)
-        self.assertLessEqual(label.thumbnail.height, Label.THUMBNAIL_HEIGHT)
-
-
-class NewLabelNonAjaxTest(ClientTest):
-
-    @classmethod
-    def setUpTestData(cls):
-        # Call the parent's setup (while still using this class as cls)
-        super(NewLabelNonAjaxTest, cls).setUpTestData()
-
-        cls.user = cls.create_user()
-
-        # Create labels and group
-        cls.create_labels(
-            cls.user, ['A', 'B'], "Group1")
-
-        cls.url = reverse('label_new')
-
-    def test_label_creation(self):
-        """Successfully create a new label."""
-        self.client.force_login(self.user)
-        self.client.post(self.url, dict(
-            name="CCC",
-            default_code='C',
-            group=LabelGroup.objects.get(code='Group1').pk,
-            description="Species C.",
-            # A new filename will be generated, and the uploaded
-            # filename will be discarded, so it doesn't matter.
-            thumbnail=sample_image_as_file('_.png'),
-        ))
-
-        # Check that the label was created, and has the expected field values
-        label = Label.objects.get(name="CCC")
-        self.assertEqual(label.default_code, 'C')
-        self.assertEqual(label.group.code, 'Group1')
-        self.assertEqual(label.description, "Species C.")
-        self.assertIsNotNone(label.thumbnail)
-        self.assertEqual(label.verified, False)
-        self.assertEqual(label.created_by_id, self.user.pk)
-
-
-class EditLabelPermissionTest(BasePermissionTest):
-
-    @classmethod
-    def setUpTestData(cls):
-        super(EditLabelPermissionTest, cls).setUpTestData()
-
-        # Create labels and group
-        labels = cls.create_labels(cls.user, ['A', 'B', 'C'], "Group1")
-
-        cls.source1 = cls.create_source(cls.user)
-        cls.create_labelset(
-            cls.user, cls.source1, labels.filter(name__in=['A', 'B']))
-        cls.source2 = cls.create_source(cls.user)
-        cls.create_labelset(
-            cls.user, cls.source2, labels.filter(name__in=['A', 'B']))
-        cls.source3 = cls.create_source(cls.user)
-        cls.create_labelset(
-            cls.user, cls.source3, labels.filter(name__in=['C']))
-
-        label_b = labels.get(name='B')
-        label_b.verified = True
-        label_b.save()
-
-        cls.url = reverse('label_edit', args=[labels.get(name='A').pk])
-        cls.url_verified = reverse('label_edit', args=[label_b.pk])
-
-    def test_anonymous(self):
-        self.assertPermissionDenied(self.url, None)
-
-    def test_admin_of_one_source_using_the_label(self):
-        user_admin_one = self.create_user()
-        self.add_source_member(
-            self.user, self.source1,
-            user_admin_one, Source.PermTypes.ADMIN.code)
-
-        # Must be admin of all sources using the label, not just one.
-        self.assertPermissionDenied(self.url, user_admin_one)
-
-    def test_editor_of_all_sources_using_the_label(self):
-        user_editor_both = self.create_user()
-        self.add_source_member(
-            self.user, self.source1,
-            user_editor_both, Source.PermTypes.EDIT.code)
-        self.add_source_member(
-            self.user, self.source2,
-            user_editor_both, Source.PermTypes.EDIT.code)
-
-        # Edit isn't enough, must be admin.
-        self.assertPermissionDenied(self.url, user_editor_both)
-
-    def test_admin_of_all_sources_using_the_label(self):
-        user_admin_both = self.create_user()
-        self.add_source_member(
-            self.user, self.source1,
-            user_admin_both, Source.PermTypes.ADMIN.code)
-        self.add_source_member(
-            self.user, self.source2,
-            user_admin_both, Source.PermTypes.ADMIN.code)
-
-        self.assertPermissionGranted(
-            self.url, user_admin_both, template='labels/label_edit.html')
-
-    def test_admin_of_all_sources_using_verified_label(self):
-        user_admin_both = self.create_user()
-        self.add_source_member(
-            self.user, self.source1,
-            user_admin_both, Source.PermTypes.ADMIN.code)
-        self.add_source_member(
-            self.user, self.source2,
-            user_admin_both, Source.PermTypes.ADMIN.code)
-
-        # Being source admin isn't enough for verified labels.
-        self.assertPermissionDenied(self.url_verified, user_admin_both)
-
-    def test_labelset_committee_member(self):
-        user_committee_member = self.create_user()
-        user_committee_member.groups.add(
-            Group.objects.get(name="Labelset Committee"))
-
-        # Committee members can edit verified labels.
-        self.assertPermissionGranted(
-            self.url_verified, user_committee_member,
-            template='labels/label_edit.html')
-
-    def test_superuser(self):
-        self.assertPermissionGranted(
-            self.url_verified, self.superuser,
-            template='labels/label_edit.html')
-
-
-class EditLabelTest(LabelTest):
-    """
-    Test label editing.
-    """
-    @classmethod
-    def setUpTestData(cls):
-        super(EditLabelTest, cls).setUpTestData()
-
-        cls.user = cls.create_user()
-
-        cls.user_committee_member = cls.create_user()
-        cls.user_committee_member.groups.add(
-            Group.objects.get(name="Labelset Committee"))
-
-        # Create labels and groups
-        cls.group1 = cls.create_label_group("Group1")
-        cls.group2 = cls.create_label_group("Group2")
-        cls.labels = dict(
-            A=cls.create_label(cls.user, "Label A", 'A', cls.group1),
-            B=cls.create_label(cls.user, "Label B", 'B', cls.group1),
-            C=cls.create_label(cls.user, "Label C", 'C', cls.group1),
-        )
-        # Ensure cls.user 'owns' the labels
-        cls.source = cls.create_source(cls.user)
-        cls.create_labelset(
-            cls.user, cls.source, Label.objects.all())
-
-    def submit_edit(self, user=None, **params):
-        # Defaults
-        post_data = dict(
-            name="Label A",
-            default_code='A',
-            group=self.labels['A'].group.pk,
-            description=self.labels['A'].description,
-        )
-        post_data.update(params)
-
-        self.client.force_login(user or self.user)
-        response = self.client.post(
-            reverse('label_edit', args=[self.labels['A'].pk]),
-            data=post_data, follow=True)
-        return response
-
-    def test_name_same(self):
-        response = self.submit_edit(name="Label A")
-
-        self.assertContains(response, "Label successfully edited.")
-
-    def test_name_change_case_only(self):
-        self.submit_edit(name="LABEL a")
-
-        self.labels['A'].refresh_from_db()
-        self.assertEqual(self.labels['A'].name, "LABEL a")
-
-    def test_name_change(self):
-        self.submit_edit(name="Label Alpha")
-
-        self.labels['A'].refresh_from_db()
-        self.assertEqual(self.labels['A'].name, "Label Alpha")
-
-    def test_name_conflict(self):
-        response = self.submit_edit(name="LABEL B")
-
-        self.assertContains(
-            response,
-            'There is already a label with the same name:'
-            ' <a href="{url}" target="_blank">'
-            '{conflicting_name}</a>'.format(
-                url=reverse('label_main', args=[self.labels['B'].pk]),
-                conflicting_name="Label B"))
-        # Check for no change
-        self.labels['A'].refresh_from_db()
-        self.assertEqual(self.labels['A'].name, "Label A")
-
-    def test_default_code_same(self):
-        response = self.submit_edit(default_code='A')
-
-        self.assertContains(response, "Label successfully edited.")
-
-    def test_default_code_change_case_only(self):
-        self.submit_edit(default_code='a')
-
-        self.labels['A'].refresh_from_db()
-        self.assertEqual(self.labels['A'].default_code, 'a')
-
-    def test_default_code_change(self):
-        self.submit_edit(default_code='Alpha')
-
-        self.labels['A'].refresh_from_db()
-        self.assertEqual(self.labels['A'].default_code, 'Alpha')
-
-    def test_default_code_conflict(self):
-        response = self.submit_edit(default_code='b')
-
-        self.assertContains(
-            response,
-            'There is already a label with the same default code:'
-            ' <a href="{url}" target="_blank">'
-            '{conflicting_code}</a>'.format(
-                url=reverse('label_main', args=[self.labels['B'].pk]),
-                conflicting_code="B"))
-        # Check for no change
-        self.labels['A'].refresh_from_db()
-        self.assertEqual(self.labels['A'].default_code, 'A')
-
-    def test_group_change(self):
-        self.submit_edit(group=self.group2.pk)
-
-        self.labels['A'].refresh_from_db()
-        self.assertEqual(self.labels['A'].group.pk, self.group2.pk)
-
-    def test_description_change(self):
-        self.submit_edit(description="Another\ndescription")
-
-        self.labels['A'].refresh_from_db()
-        self.assertEqual(self.labels['A'].description, "Another\ndescription")
-
-    def test_thumbnail_change(self):
-        original_filename = self.labels['A'].thumbnail.name
-
-        self.submit_edit(thumbnail=sample_image_as_file('_.png'))
-
-        # Check for a different thumbnail file, by checking filename.
-        # This assumes the filenames are designed to not clash
-        # (e.g. qoxibnwke9.jpg) rather than replace each other
-        # (e.g. thumbnail-for-label-29.jpg).
-        self.labels['A'].refresh_from_db()
-        self.assertNotEqual(original_filename, self.labels['A'].thumbnail.name)
-
-    def test_verified_change(self):
-        self.submit_edit(
-            user=self.user_committee_member,
-            verified=True)
-
-        self.labels['A'].refresh_from_db()
-        self.assertEqual(self.labels['A'].verified, True)
-
-    def test_verified_requires_permission(self):
-        # Non committee member
-        self.submit_edit(
-            user=self.user,
-            verified=True)
-
-        self.labels['A'].refresh_from_db()
-        # Not changed
-        self.assertEqual(self.labels['A'].verified, False)
-
-    def test_duplicate_change(self):
-        # Must verify B to make it a candidate for a duplicate pointer
-        self.labels['B'].verified = True
-        self.labels['B'].save()
-
-        response = self.submit_edit(
-            user=self.user_committee_member,
-            duplicate=self.labels['B'].pk)
-
-        self.assertContains(response, "Label successfully edited.")
-        self.labels['A'].refresh_from_db()
-        self.assertEqual(self.labels['A'].duplicate.pk, self.labels['B'].pk)
-
-    def test_duplicate_requires_permission(self):
-        # Must verify B to make it a candidate for a duplicate pointer
-        self.labels['B'].verified = True
-        self.labels['B'].save()
-
-        response = self.submit_edit(
-            user=self.user,
-            duplicate=self.labels['B'].pk)
-
-        # Still expect to show success, but with the duplicate field ignored
-        self.assertContains(response, "Label successfully edited.")
-
-        self.labels['A'].refresh_from_db()
-        # Not changed
-        self.assertEqual(self.labels['A'].duplicate, None)
-
-    def test_duplicate_cant_point_to_unverified(self):
-        self.labels['B'].verified = False
-        self.labels['B'].save()
-
-        response = self.submit_edit(
-            user=self.user_committee_member,
-            duplicate=self.labels['B'].pk)
-
-        self.assertContains(response, "Please correct the errors below.")
-        self.assertContains(
-            response,
-            "Select a valid choice. That choice is not one of the available"
-            " choices.")
-
-        self.labels['A'].refresh_from_db()
-        # Not changed
-        self.assertEqual(self.labels['A'].duplicate, None)
-
-    def test_duplicate_cant_be_verified(self):
-        self.labels['B'].verified = True
-        self.labels['B'].save()
-
-        response = self.submit_edit(
-            user=self.user_committee_member,
-            verified=True,
-            duplicate=self.labels['B'].pk)
-
-        self.assertContains(response, "Please correct the errors below.")
-        self.assertContains(
-            response, "A label can not both be a Duplicate and Verified.")
-
-        self.labels['A'].refresh_from_db()
-        # Not changed
-        self.assertEqual(self.labels['A'].duplicate, None)
-
-
-class PopularityTest(ClientTest):
-    """Tests related to label popularity values."""
-
-    @classmethod
-    def setUpTestData(cls):
-        super(PopularityTest, cls).setUpTestData()
-        cls.user = cls.create_user()
-        cls.source = cls.create_source(
-            cls.user,
-            point_generation_type=PointGen.Types.SIMPLE,
-            simple_number_of_points=2)
-
-        cls.labels = cls.create_labels(
-            cls.user, ['A', 'B'], "Group1")
-        cls.label_a = cls.labels.get(name='A')
-
-        cls.img = cls.upload_image(cls.user, cls.source)
-
-    def setUp(self):
-        super(PopularityTest, self).setUp()
-
-        # Popularities are cached when computed, so we clear the cache to
-        # prevent a previous test from affecting the next one.
-        cache.clear()
-
-    def test_zero_sources(self):
-        # There's a labelset, but it doesn't have A
-        self.create_labelset(
-            self.user, self.source, self.labels.filter(name='B'))
-
-        self.assertEqual(
-            self.label_a.popularity, 0,
-            msg="0 sources should mean 0 popularity")
-
-    def test_zero_annotations(self):
-        # A is in a labelset
-        self.create_labelset(self.user, self.source, self.labels)
-        # There are annotations, but they're not of A
-        self.add_annotations(self.user, self.img, {1: 'B'})
-
-        self.assertEqual(
-            self.label_a.popularity, 0,
-            msg="1 source and 0 annotations still should mean 0 popularity")
-
-    def test_nonzero_annotations(self):
-        # A is in a labelset
-        self.create_labelset(self.user, self.source, self.labels)
-        # A has annotations (by a quirk of the formula, it actually needs more
-        # than 1 annotation to get non-0 popularity)
-        self.add_annotations(self.user, self.img, {1: 'A', 2: 'A'})
-
-        self.assertGreater(
-            self.label_a.popularity, 0,
-            msg="Non-0 annotations should mean non-0 popularity")
-
-    def test_popularity_cached(self):
-        # A is in a labelset
-        self.create_labelset(self.user, self.source, self.labels)
-
-        self.assertEqual(
-            self.label_a.popularity, 0,
-            msg="Popularity should be 0")
-
-        # Now A has annotations, and would have non-0 popularity if the
-        # cache refreshed
-        self.add_annotations(self.user, self.img, {1: 'A', 2: 'A'})
-
-        self.assertEqual(
-            self.label_a.popularity, 0,
-            msg="Cached popularity of 0 should still be used")
